+++ src/DatatableCrudsProvider.php
@@ -2,8 +2,10 @@
 
 namespace Exist404\DatatableCruds;
 
-use Exist404\DatatableCruds\Console\Commands\PrepareCommand;
 use Exist404\DatatableCruds\Console\Commands\InstallPackageCommand;
+use Exist404\DatatableCruds\Middleware\DatatableInjection;
+use Illuminate\Contracts\Http\Kernel;
+use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\ServiceProvider;
 
 class DatatableCrudsProvider extends ServiceProvider
@@ -29,9 +31,16 @@ class DatatableCrudsProvider extends ServiceProvider
         $this->loadViewsFrom(__DIR__ . '/views', 'datatable');
         $this->loadRoutesFrom(__DIR__ . '/routes/web.php');
         $this->publishes([__DIR__ . '/../config/datatablecruds.php' => config_path('datatablecruds.php')], 'config');
-        $this->commands([
-            InstallPackageCommand::class,
-            PrepareCommand::class,
-        ]);
+        $this->commands([InstallPackageCommand::class]);
+        $this->registerMiddleware(DatatableInjection::class);
+        Blade::directive('datatable', function ($data) {
+            return "<data-list :data='{{ json_encode($data) }}'></data-list>";
+        });
+    }
+
+    protected function registerMiddleware($middleware)
+    {
+        $kernel = $this->app[Kernel::class];
+        $kernel->pushMiddleware($middleware);
     }
 }