+++ src/helper.php
@@ -1,7 +1,10 @@
 <?php
 
+use Illuminate\Contracts\Pagination\LengthAwarePaginator;
+use Illuminate\Database\Eloquent\Collection;
+
 if (!function_exists('dataTableOf')) {
-    function dataTableOf($model = null)
+    function dataTableOf($model = null): LengthAwarePaginator|Collection
     {
         return (new \Exist404\DatatableCruds\ModelDataTable($model))->get();
     }