+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Release Notes
+## [v1.0.9](https://github.com/404Exist/datatable-cruds/compare/v1.0.8...v1.0.9) - 2022-12-10
+
+### Added
+
+- Ability to render multiple datatables in same page
+- `renderData()` method have been added.
+- `@datatable()` directive have been added.
+
 ## [v1.0.8](https://github.com/404Exist/datatable-cruds/compare/v1.0.7...v1.0.8) - 2022-09-23
 
 ### Added